Minor reorganization in ImDebugger
