+++ UI/ImDebugger/ImDebugger.h
@@ -31,76 +31,6 @@ class MIPSDebugInterface;
 class GPUDebugInterface;
 struct ImConfig;
 
-// Corresponds to the CDisasm dialog
-class ImDisasmWindow {
-public:
-	void Draw(MIPSDebugInterface *mipsDebug, ImConfig &cfg, ImControl &control, CoreState coreState);
-	ImDisasmView &View() {
-		return disasmView_;
-	}
-	void NotifyStep() {
-		disasmView_.NotifyStep();
-	}
-	void DirtySymbolMap() {
-		symsDirty_ = true;
-	}
-	const char *Title() const {
-		return "CPU Debugger";
-	}
-
-private:
-	// We just keep the state directly in the window. Can refactor later.
-
-	enum {
-		INVALID_ADDR = 0xFFFFFFFF,
-	};
-
-	u32 gotoAddr_ = 0x08800000;
-
-	// Symbol cache
-	std::vector<SymbolEntry> symCache_;
-	bool symsDirty_ = true;
-	int selectedSymbol_ = -1;
-	char selectedSymbolName_[128];
-
-	ImDisasmView disasmView_;
-	char searchTerm_[64]{};
-};
-
-// Corresponds to the CMemView dialog
-class ImMemWindow {
-public:
-	void Draw(MIPSDebugInterface *mipsDebug, ImConfig &cfg, ImControl &control, int index);
-	ImMemView &View() {
-		return memView_;
-	}
-	void DirtySymbolMap() {
-		symsDirty_ = true;
-	}
-	void GotoAddr(u32 addr) {
-		gotoAddr_ = addr;
-		memView_.gotoAddr(addr);
-	}
-	static const char *Title(int index);
-
-private:
-	// We just keep the state directly in the window. Can refactor later.
-	enum {
-		INVALID_ADDR = 0xFFFFFFFF,
-	};
-
-	// Symbol cache
-	std::vector<SymbolEntry> symCache_;
-	bool symsDirty_ = true;
-	int selectedSymbol_ = -1;
-	char selectedSymbolName_[128];
-
-	ImMemView memView_;
-	char searchTerm_[64]{};
-
-	u32 gotoAddr_ = 0x08800000;
-};
-
 // Snapshot of the MIPS CPU and other things we want to show diffs off.
 struct ImSnapshotState {
 	u32 gpr[32];
